media-keys: Stop treating tablets specially
There are problems with treating the power-button on tablets in a special
manner:

1) The configuration for the power-button behavior is now part of the
main control-center power capplet, so it is visible to all users. It is
very strange and inconsistent for users with a device which gets identified
as a tablet that they can configure the behavior of the button in the
control-center but the behavior does not change.

2) The method used to determine if something is a tablet is unreliable,
various surface like devices and some laptops wrongly have "Tablet"
as chassis-type. So relying on chassis-type info for this is a bad idea.

This commit removes the special handling of the power-button on tablets,
falling back to the normal behavior of doing what the user has configured.