color: Always use old blackbody calculation
We tried to switch to a better way of calculating the blackbody color,
but unfortunately the switch was only partial and not used in most
cases. Revert this to the old calculation for stable so that we can
update the default values and configuration panel when moving to the
better blackbody color calculation.

Fixes #403