color: Switch to better black body color calculation
The planckian calculation is much better, but it also means we need to
change the default from 4000K to 2700K to get a similar effect. 2700K is
actually a good value as it is similar to commonly found "warm-white"
artificial lights.

Note that the defaults and range in g-c-c will also need to be adjusted
for this change.