media-keys: Fetch device node from AcceleratorActivated arguments
Gnome-shell now passes the additional device-node parameter in the variant
we can use to directly look up the right stream for the key event, so
pass it all the way through.

In order to add some wiggle room, still handle the device-id argument as
a fallback, so it works in combination with older gnome-shell. Only do it
on X11 sessions though, as the device ID is only meaningful there.

Related: https://gitlab.gnome.org/GNOME/gnome-shell/merge_requests/618