power: Fix rounding of brightness value
We rounded properly for percentage to absolute value, but we always
floored the other way around. Also round when converting the absolute
value to percentage, this should prevent issues where the percentage
changes becaues the calculated absolute value will be resolved to a
different percentage than it was calculated from.