Use atime instead of mtime in thumbnail cleaner
Currently we expire thumbnails some amount of time after creation, even
if the thumbnail is regularly accessed. This is probably not what we
want to happen. Using atime instead fixes this behavior.

If the user is running on a file system with atime disabled, this is
still no worse than using mtime.