xsettings: Outsource decision making regarding animations to the shell
Animations were disabled depending various hueristics, specifically
whether there was a Vino session active on D-Bus, whether the desktop
was hardware accelerated, or if a certain X11 extension was advertised
that happens to be advertised when Xvnc is used as an X server.

This logic has been moved to gnome-shell, and is exposed as a property
on org.gnome.Shell.Introspect, except Vino sessions is not discovered
anymore. gnome-shell has, however, means to detect whether a remote
desktop session that wants to have animations disabled is active or not.