media-keys: Let the screenshot service save to clipboard itself
This is both broken currently on wayland (since this daemon doesn't
request the x11 backend, and wayland clients need a proper surface/serial
to request clipboard), and conceptually wrong (we wouldn't like to
enable the X11 backend just to let it save to clipboard).

Let gnome-shell handle saving the screenshot to clipboard instead, this
also saves us from saving to /tmp just to copy it to the clipboard
afterwards. We achieve this by passing an empty string as the filename.

Related: https://gitlab.gnome.org/GNOME/mutter/issues/789
         https://gitlab.gnome.org/GNOME/gnome-shell/merge_requests/810