media-keys: Use X11 display for clipboard
This daemon does not specify to be X11, and arguably "make it able
to save to clipboard" is hardly a reason. Make it use a X11 display
specifically for saving screenshots to clipboard, as it's got at
least some guarantees of working there.

Related: https://gitlab.gnome.org/GNOME/mutter/issues/789