power: Use logind helper to change backlight brightness
This is the favored method, over our own polkit helper, and over RandR.
Since this helper does its own smart queueing of events, just forward them
as fast as possible, and mark all tasks as completed ASAP on our side.

However, as this is a recent logind feature (New in systemd 243), make this
a soft runtime dependency, by checking first that the required DBus method
exists. We forget about logind and go through the other methods if this one
is not available.

Closes: https://gitlab.gnome.org/GNOME/gnome-settings-daemon/issues/429