media-keys: Place spawned processes into a systemd scope
This means we isolate the processes from the media-keys scope itself.
The scope units are created with a gnome-launched- prefix, which allows
us to create a GNOME wide drop-in to set defaults.

If systemd support is disabled, or we are not running on a systemd
session, then the code will do nothing.