dummy: Do not start gsd-dummy for stub desktop files
We install stub desktop files so that session can pull these in as
required components. However, for gnome-session to be happy, we do not
actually need to start a process. It is sufficient to install the
.desktop file and then disable the service using Hidden=true.

This means we can replace the Exec line with e.g. /bin/false, removing
the need for the gsd-dummy binary.