meson: Move stub desktop file generation into plugins loop
There is no need to have two places for desktop file generation for
services that can be disabled. Instead, integrate the logic into the
main plugin loop and simply use a different template file as the input.

To implement this, create a list of disabled modules that can be tested
to skip the generation of the systemd units and inclusion of the plugin
directory.