backlight: Fix return value of async brightness changes
With cd051e6c7ccfb7e7b0ec440409c559b2e51284e8 we added another method
for setting the display brightness, it was forgotten there to convert
the actual brightness value to a percentage when setting the GTask
return value.

Since this return value is used to display the brightness OSD in
gsd-media-keys-manager, we ended up with values above 100% shown in the
brightness OSD. Fix that by adjusting the return value to be a
percentage.