meson: Rename plugin_install_wants to plugin_gate_units
The idea is to generate a gsd-X-gate.target intermediate unit which
has a requisite on all units listed in plugin_gate_units.

i.e. for most services we are going to have:
 * gsd-X.target: marker on whether services should run in the session
 * gsd-X.service: actual service that is started

This changes for units that should only run under certain conditions
(e.g. X11 or smartcard). In that case we add
 + gsd-X-gate.target
which has a proper requisite on all required units and ensures that the
servie will only start if all dependencies are met. This gate unit is
primarily needed so that we can still have an OnFailure target in the
.service file without issues.