plugins: Update systemd unit design and generate them
Rather than having per-plugin units, generate them from the meson build
file using a common template.

The unit layout is changed somewhat. The .target file now is solely a
flag on whether the service is requested for the session type. The
variosu requirements are all moved into the .service file.

Instead of using an OnFailure and trying to contain the dependency
errors in a separate unit, we now use ExecStopPost. It seems that the
current approach has never worked anyway.

The Conflict entries on certain session types are gone now. This is
instead handled outside of gnome-settings-daemon.