usb: always allow hardwired devices rather than each and every device
This change makes use of a sysfs feature through USBGuard, namely the
connect-type. USBGuard exposes it since
https://github.com/USBGuard/usbguard/pull/274 got merged. connect-type
comes from sysfs and tries to tell whether a port is fixed or removable.
It takes that information from ACPI, so the firmware needs to be well
written. Some vendors (e.g. Dell) have correct information while others
(e.g. Lenovo) only mark external ports as "hotplug" and do not
explicitly mark internal ports as "hardwired".
This change changes existing functionality in that is does not allow all
devices any more. So we will need to add that functionality back.
This change is meant to be a safe-guard to protect the user from being
locked out accidentally. We assume that devices marked as internal (by
the firmware) are more trustworthy than hot-pluggable devices.