usb: leave a comment on the volatility of the blocking rule
The blocking rule is what might break our neck since it causes the USB
devices to not work. We must exercise great care to not overblock. And
if it was possible to create temporary policies, we might be better off.
This comments reflects the thinking of the module and might serve as a
motiviation for USBGuard to implement such a behaviour.