usb: Detect whether a device has been allowed by USBGuard
We listen to DevicePolicyChanged instead of DevicePresenceChanged.
This allows for knowing whether USBGuard has blocked (or allowed) a
device.  We intended to do that before, but the signal used before
(DevicePresenceChanged) is sent very early in the process by USBGuard.
In fact, the documentation states that it's not very interesting for 3rd
party apps to consume that signal. And I concur.
With the signal now in use, we get the rule that has been applied to the
device. And hence, we get to know whether the device has been allowed by
USBGuard.