usb: detect and allow hardwired devices
This change intends to make sure that internal devices always work.
It is meant to be a safe-guard to protect the user from being
locked out accidentally. We assume that devices marked as internal (by
the firmware) are more trustworthy than hot-pluggable devices.

It makes use of a sysfs feature through USBGuard, namely the
connect-type. USBGuard exposes it since
https://github.com/USBGuard/usbguard/pull/274 got merged. connect-type
comes from sysfs and tries to tell whether a port is fixed or removable.
It takes that information from ACPI, so the firmware needs to be well
written. Some vendors (e.g. Dell) have correct information while others
(e.g. Lenovo) only mark external ports as "hotplug" and do not
explicitly mark internal ports as "hardwired".

Note though: USBGuard actually does not provide that information in a
structured way yet. But chances are very high that it will happen and
then we can immediately make use of that property.