usb: prepend the allow rule as temporary
We shouldn't write permanent rules in the usbguard unless necessary.
In our case we relay on the allow rule to authorize USB devices when the
lockscreen is off.

If something goes wrong, after a reboot we expect that the user's
devices will be authorized by a default ImplicitPolicyTarget "allow".
Note however that this is still not the case for the upstream usbguard.