Use DevicePresenceChanged for detecting a plugged in device
We have tried to not use this signal, because it is "of little use".
There is, however, no other way for us to get notified about a device
being plugged in.

The downside is that we don't know whether the device will be allowed.
This is a bit silly, but we can only for upstream making the change
soon.