Add capability to disable USB while the lockscreen is on
This change adds the ability to make use of UBSGuard to disable new USB
devices while the lockscreen is on. Note that USBGuard as of 0.7.5 is
needed.

In order to make use of this feature, you need to enable it first,
e,g, through gsettings set org.gnome.desktop.privacy usb-protection true.
And you need to have USBGuard installed, of course. Beware though: With the
default configuration the user is likely to be locked out of their
system. We encourage downstreams to change the default configuration
value s.t. USB devices are enabled unless a rule explicitly blocks them.
We take some care about users in that we allow keyboards and hardwired
devices, though.

The intended behaviour is: If the lockscreen is on, newly inserted USB
devices will not be accepted. A notification is given to the user that
they have to unlock the session and re-plug the device.

Ludovico de Nittis <denittis@gnome.org> did most of the work.