housekeeping: Never cross into a different filesystem
This stops us from crossing into another file system. The test simply
uses the file system tag, which means that e.g. bind mounts to the same
file system are not found.

The exception to this is trash:, in that case, set filesystem to NULL
and allow crossing over. This must be done as trash: is virtual and
files might be coming from various locations.

Fixes: #252