datetime: reduce runtime memory usage
Just over 1 MB of runtime data was keeping track of allocations for
TzLocation and containing GList elements. We can make much of that all go
away if we use intern'd strings and arrays to contain TzLocation data.

A much bigger amount of allocations are held by libgweather which is
storing rougly 8MB of Locations.xml data in memory. A follow up commit
will try to address that more directly.