datetime: use malloc_trim() to release resources back to OS
We allocate a lot of memory while building the timezone database from
libgweather all to have that put into the GSlice allocation queues we'll
barely if ever touch again.

By forcing things to use malloc, we can use malloc_trim() on supported
GLibc systems to madvise() the kernel it can take the memory back.

With the previous patches and this, gsd-datetime drops from about 9MB down
to just under 3MB. We could probably go even further if we change
libgweather a bit, but this gets it small enough for me.

However, there are other processes (such as evolution-calendar-factory)
that would benefit from a more generic solution.

It's also very likely that we would want to just make gsd-datetime
activatable through external events so the process can go away entirely.