media-keys: Ignore our own output when omitting sound feedback
Since commit 7d74327dc2b we omit sound feedback on volume changes
if something is already outputting sound. Unfortunately that
"something" may be our own feedback (from a previous volume
change).

We do not want to omit the feedback in that case, so only consider
sound playing if something *else* is outputting sound.

https://gitlab.gnome.org/GNOME/gnome-settings-daemon/-/issues/508