media-keys: Do not suppress volume-change while playing
This effectively reverts commit 7d74327dc2 ("media-keys: Skip
volume-change feedback while playing"). The main problem is that we get
into our own way when the volume-change happens multiple times in quick
succession.

The feature is nice, but we need to improve it so that playback is only
supressed when a proper audio stream is playing rather than just system
sounds.

Resolves: #508