media-keys: Add static mapping for XF86RotationLockToggle
Some tablets have a dedicated button for toggling accelerometer based
rotation on/off. A while ago the XF86RotationLockToggle keysym was added
for this and this has been hooked up through all the plumbing layers
(xkeyboard-config, libxkb-common) now.

Add a static mapping for this, to make these buttons work ootb.
This has been tested with such a button on a Lenovo Thinkpad 10 tablet.