power: Only install callback once
A mode change to normal causes iio_proxy_claim_light() to be called.
Prior to this patch, this would always connect iio_proxy_changed_cb(),
even if the previous mode was dim and the callback was already
connected.  This resulted in the callback triggering multiple times on
each change of the ambient light sensor.  This patch adds a flag to
GsdPowerManager to track whether or not the callback is already
connected.