power: Only install callback once
A mode change to normal causes iio_proxy_claim_light() to be called with
active set to TRUE.  This connects iio_proxy_changed_cb() even if the
previous mode was dim and the callback was already connected.  This
caused the callback to be triggered multiple times on each change of the
ambient light sensor.  With this patch, calls to iio_proxy_claim_light()
always remove the callback first, and if active is TRUE they add it
back.