meson: Fix masking symlink location and detection of jhbuild
The symlinks need to be installed into DESTDIR, use a shell to do this
correctly. Also, doing this on debug builds is not very helpful,
instead, do it by testing whether we are installing into the same prefix
that systemd is coming from.

This assumes that we will run using the same systemd instance as we are
picking up the pkgconfig file from, but that seems like a reasonable
assumption (i.e. it is unlikely anyone has systemd in their prefix and
doesn't bother to set things up correctly).