usb-protection: Relax the whitelist to all HIDs and HUBs
We discriminate keyboards when the screen is locked, i.e. we allow them
so that the user can input their password.
But when the keyboard is behind a hub, e.g. the laptop on a docking
station, then the keyboard does not work.

The mouse is probably also commonly used to unlock the screen before
typing the password.

Because this protection is enabled by default, and we don't want to get
in the way of the user and their session, with this commit we expand the
class of allowed devices to all HIDs and HUBs.

Stricter protection features might be evaluated in the future.