xsettings: Fix race condition in fontconfig timestamp test
gsd-xsettings needs some time to update the fontconfig caches before it
also updates the corresponding timestamp in the X property. So add
delay logic that first waits for the changes to be picked up and
processed, and then waits a bit longer just to be sure we wait long
enough.