usb: defensively destructure GVariants to prevent crashes
The data is coming in from the outside, so it is untrusted. We check for
the size of the variant, because otherwise we might crash at runtime,
since g_variant_get_child might leave variables, i.e. GVariantIters,
uninitialised (NULL) which then causes g_variant_iter_loop to crash,
since it doesn't check for NULL.

The check for g_variant_n_children is arguably ugly, because
g_variant_get_child will eventually perform that check.
So we focus on the returned NULL in case of g_variant_get_child_value
or, in case of g_variant_get_child, whether the memory has not changed.

The documentation is, for my taste, not as clear as it could or should be.
It mentions that it is "an error" to provide an index higher than what
the GVariant contains. But it doesn't mention how it communicates that
error nor how to deal with it.