usb: authorize device after screen has been locked
This should make it a bit more safe, because the device will not be
authorized before the screen is actually locked. Before, the device
would be authorized and then it asks for the screen to be locked.
This leaves a small window where the device could interact with the
unlocked session. I don't think it is relevant in practice, though.
Anyway, it is semantically more correct to authorize after the screen
has been locked.

This change brings an unfortunate indirection of arguments,
i.e. you have to place the id in the manager structure.
This feels clunky and should be revised.
We should also handle the failure case a bit more gracefully, I think. For
example, show a notification about the device that we wanted to
authorize, but couldn't.