plugins: Use Restart=on-failure even for XSettings plugin
The move to use Restart=on-abnormal happened because gsd-xsettings would
die if Xwayland shuts down. However, since gnome-shell commit 01a927f388
and related changes, we should be waiting for gsd-xsettings to shut down
gracefully before Xwayland is stopped.

As such, we should only hit an exit failure if from gsd-xsettings if
Xwayland crashes. If gnome-shell follows up with a stop of the services
within 100ms, then such a restart will not even be attempted.