media-keys: Don't plonk on screenshot errors
From a user's point of view, it doesn't matter whether gnome-shell's
screenshot API returned `false` or a D-Bus error: Both mean that
a screenshot was not taken.

Treating them differently is therefore slightly confusing, so stop
doing that: Rely on the lack of flash effect and shutter sound as
feedback and remove the "plonk" sound played after errors.