xsettings: Add support for XFixes ClientDisconnectMode
The Xserver itself is capable of terminating itself once all X11 clients
are gone, yet in a typical full session, there are a number of X11
clients such as gsd-xsettings running continuously.

Those always-running clients will prevent the Xserver from terminating,
because the actual number of X11 clients will never drop to 0.

Use XFixes ClientDisconnectMode to inform the X11 server that it can
terminate even if gsd-xsettings is still running.

That will allow Xwayland to terminate automatically once regular clients
have quit.

On plain Xorg servers, the lifetime of the session is usually tied to
the session manager or window manager, and this change will have no
effect.