rfkill: Do not sync state immediately when opening rfkill
We are opening the device asynchronously anyway. There is really no need
to sync the state immediately, we can simpl rely on the POLL handler to
do so later on from the mainloop.

The difference we should see is that the debug information about there
being no rfkill devices is gone. And, that some of the debug messages
will be slightly different.