tests: Wait longer for brightness updates
Sometimes the test_legacy_brightness_rounding seems to fail, in rare
cases with the brightness file being empty. This is possible in busy CI
environments due to the short time window (<= 200ms) and because the
script first truncates the file.

Increase the timeout in order to avoid such a CI failure. Note that
writing the update atomically is not really feasible.