power: Dim screen faster if power saver mode is on
As done on other platforms, aggressively dim the screen after a
short period when the user has selected to enter power saver mode.

The same aggressive screen dim will be used if the battery is low and
power-profiles-daemon is not available. If it is available, then it
fixes a screen dim happening when the battery was low which might
have been unwanted.

See https://gitlab.gnome.org/GNOME/gnome-control-center/-/issues/232

Prior art:
https://support.apple.com/en-us/HT205234