power: Add comment about skipped types in low battery notifications
Those device types that exist in upower (and gnome-bluetooth) don't have
a user-friendly precise name that we could use to show in the
notification so fallback to the generic name instead.

eg. a "Wearable" could be a watch, or a number of other devices, but
end-users don't know what a "wearable" is in Bluetooth parlance.