color/night-light: Make sure to set target temperature
When smoothing a temperature, we practically never ended up setting the
target temperature, since the throttling dropped the last few changes.
For example, after night light was disabled, the temperature tended to
get stuck on 6494 instead of 6500. Fix this by making sure the target
temperature is forced, ignoring the throttling.

Also disable throttling when not smoothing, as it's likely we want the
temperature that is actually asked for.