xsettings: Adopt code to look up GTK IM module
Right now, gsd-keyboard and gsd-xsettings have a strange relation
where the first sets the gtk-im-module dconf setting for the latter
to read the setting and forward it through XSettings.

Since this detection is highly X11 specific, make it happen in the
Xsettings daemon itself, from the relevant setting and device presence.

This makes users still able to shoot themselves in the foot by changing
the setting, X11 clients able to be told to switch to ibus if it turns
out necessary, and Wayland clients unaffected otherwise.

Related: https://gitlab.gnome.org/GNOME/gtk/-/issues/4443