tests: Enable dconf service for new dbusmock compatibility
python-dbusmock will also stop importing all session services that are
installed on the host system by default (as is already the case for
system services). With this change, the dconf service cannot be
activated anymore unless we explicitly enable the service in the
environment.

Add the appropriate call to enable the service, guarded by a check that
the method exists to be compatible with older dbusmock versions.

See: https://github.com/martinpitt/python-dbusmock/pull/108