usb: Use more accurate DevicePolicySignal to determine status of a device
The signal is new and was introduced to allowed for correct assessment of
whether a device has been authorised or blocked.
The DevicePresenceChanged signal would not tell which rule finally applied
to a device. So a device could be rejected initially but allowed by
configuration. In that case, a PolicyChanged signal will be thrown just after.
But we would need to wait for that signal to determine whether a device will be allowed.
And we don't know for how long we would need to wait.
This signal tells consumers the result of all configuration.

This requires USBGuard >= 1.1.0 (2022-02-24) with this patch:
https://github.com/USBGuard/usbguard/pull/459