usb: use logind's LockedHint to accurately reflect the locked state of the session
Before this change, we tracked the status of session lock through the
screensaver proxy. But it's not ideal, because it fires in the wrong
moment. More precise is LockedHint.

The problem manifests itself when locking the screen, e.g. with Ctrl+L
and moving the mouse. The screen would be locked, but the signal will
not be thrown.

https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/351

First, I tried to use cached values like so:

    g_autoptr(GVariant) cached_reply = g_dbus_proxy_get_cached_property
    (logind_proxy, "LockedHint");
        if (cached_reply) {
            result = g_variant_get_boolean (cached_reply);
        } else {
            actually call dbus
        }

but it didn't detect when I had the screen locked. It still used the old
value indicating that the session was unlocked.
I thus decided to not use caching at all.
It should probaby just work™, but somehow it doesn't.