usb: address race when authorising devices
There could be a pathological case where locking of the screen takes
very long. If a new device is inserted while the other is not yet
authorised, we would only authorise the new and not the old device.
That's undesirable, because those devices are HIDs which we are trying
to allow to enable users to recover from broken keyboards.

With this change, we maintain less global state which arguably makes it
easier to reason about the code. The need for allocating memory is
unfortunate, but I don't know a better way. I would love to simply pass
a second argument to the callback.

We cannot cancel ongoing authorisations. But it also doesn't seem like a
problem. The device IDs to be authorised hardly repeat. So far they are
incremental. It could be possible to overflow and roll over. But the IDs
are UINT32. So an attacker would have to re-plug a device about four
billion times during a relatively short period.