housekeeping: clean up x-large and xx-large thumbnails
Support for two new thumbnails directories named "x-large" and
"xx-large" were added to the XDG thumbnail-spec in 2020 and to
gnome-desktop in GNOME 40. They started to be used by nautilus in GNOME
43 to have larger/sharper thumbnails on HiDPI screens.

But the housekeeper never got updated, so the thumbnail cache has
unbounded growth. I realized this problem when I was trying to debug a
low disk space condition, and realized that my thumbnail cache was 10GiB
in size.

Add the new directories. They're only added to the XDG cache path, since
any application new enough to generate these thumbnail sizes would be
using that path.